API-205_API-161: Extract `executeStreamRequest` function from each partialUpdateList tests classes
