+++ src/Pim/Bundle/ApiBundle/tests/integration/ApiTestCase.php
@@ -6,6 +6,7 @@ use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\ConnectionCloser;
 use Akeneo\Test\Integration\DatabaseSchemaHandler;
 use Akeneo\Test\Integration\FixturesLoader;
+use Pim\Bundle\ApiBundle\Stream\StreamResourceResponse;
 use Symfony\Bundle\FrameworkBundle\Client;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
@@ -238,5 +239,53 @@ abstract class ApiTestCase extends WebTestCase
         }
 
         $esClient->createIndex($conf->buildAggregated());
+    }
+
+    /**
+     * Execute a request where the response is streamed by chunk.
+     *
+     * The whole content of the request and the whole content of the response
+     * are loaded in memory.
+     * Therefore, do not use this function with an high input/output volumetry.
+     *
+     * @param string $method
+     * @param string $uri
+     * @param array  $parameters
+     * @param array  $files
+     * @param array  $server
+     * @param string $content
+     * @param bool   $changeHistory
+     *
+     * @return array
+     */
+    protected function executeStreamRequest(
+        $method,
+        $uri,
+        array $parameters = [],
+        array $files = [],
+        array $server = [],
+        $content = null,
+        $changeHistory = true
+    ) {
+        $streamedContent = '';
+
+        ob_start(function ($buffer) use (&$streamedContent) {
+            $streamedContent .= $buffer;
+
+            return '';
+        });
+
+        $client = $this->createAuthenticatedClient();
+        $client->setServerParameter('CONTENT_TYPE', StreamResourceResponse::CONTENT_TYPE);
+        $client->request($method, $uri, $parameters, $files, $server, $content, $changeHistory);
+
+        ob_end_flush();
+
+        $response = [
+            'http_response' => $client->getResponse(),
+            'content'       => $streamedContent,
+        ];
+
+        return $response;
     }
 }
