+++ src/Pim/Bundle/ApiBundle/tests/integration/Controller/Channel/PartialUpdateChannelIntegration.php
@@ -310,9 +310,6 @@ JSON;
      */
     protected function getConfiguration()
     {
-        return new Configuration(
-            [Configuration::getTechnicalCatalogPath()],
-            true
-        );
+        return new Configuration([Configuration::getTechnicalCatalogPath()]);
     }
 }
