+++ src/Pim/Bundle/ApiBundle/tests/integration/Controller/Family/PartialUpdateListFamilyIntegration.php
@@ -284,54 +284,6 @@ JSON;
     }
 
     /**
-     * Execute a request where the response is streamed by chunk.
-     *
-     * The whole content of the request and the whole content of the response
-     * are loaded in memory.
-     * Therefore, do not use this function on with an high input/output volumetry.
-     *
-     * @param string $method
-     * @param string $uri
-     * @param array  $parameters
-     * @param array  $files
-     * @param array  $server
-     * @param string $content
-     * @param bool   $changeHistory
-     *
-     * @return array
-     */
-    protected function executeStreamRequest(
-        $method,
-        $uri,
-        array $parameters = [],
-        array $files = [],
-        array $server = [],
-        $content = null,
-        $changeHistory = true
-    ) {
-        $streamedContent = '';
-
-        ob_start(function($buffer) use (&$streamedContent) {
-            $streamedContent .= $buffer;
-
-            return '';
-        });
-
-        $client = $this->createAuthenticatedClient();
-        $client->setServerParameter('CONTENT_TYPE', StreamResourceResponse::CONTENT_TYPE);
-        $client->request($method, $uri, $parameters, $files, $server, $content, $changeHistory);
-
-        ob_end_flush();
-
-        $response = [
-            'http_response' => $client->getResponse(),
-            'content'       => $streamedContent,
-        ];
-
-        return $response;
-    }
-
-    /**
      * @param array  $expectedFamily normalized data of the family that should be created
      * @param string $code           code of the family that should be created
      */
